Template constructor of DistributedSearchTree on the primitives
